Move tuple-array Python test to JVM-only source set
